+++ src/GongwenPatcher/lib/Extensibility.cs
@@ -1,0 +1,32 @@
+using System;
+
+namespace Extensibility
+{
+    public enum ext_ConnectMode
+    {
+        ext_cm_AfterStartup = 0,
+        ext_cm_Startup = 1,
+        ext_cm_External = 2,
+        ext_cm_CommandLine = 3,
+        ext_cm_Solution = 4,
+        ext_cm_UISetup = 5
+    }
+
+    public enum ext_DisconnectMode
+    {
+        ext_dm_HostShutdown = 0,
+        ext_dm_UserClosed = 1,
+        ext_dm_UISetupComplete = 2,
+        ext_dm_SolutionClosed = 3
+    }
+
+    [System.Runtime.InteropServices.ComVisible(true)]
+    public interface IDTExtensibility2
+    {
+        void OnConnection(object Application, ext_ConnectMode ConnectMode, object AddInInst, ref Array custom);
+        void OnDisconnection(ext_DisconnectMode RemoveMode, ref Array custom);
+        void OnAddInsUpdate(ref Array custom);
+        void OnStartupComplete(ref Array custom);
+        void OnBeginShutdown(ref Array custom);
+    }
+}